refactor(import): split FileImportService and extract helper views/services
- Extract ImportEnrichmentService and its worker threads into separate files to handle metadata enrichment and downloads.
- Relocate DuplicateImportDialog and BatchImportProgressDialog views out of FileImportService.swift.
- Maintain FileImportService as the main import facade/coordinator without changing transactional, commit/rollback, or repository storage logic.
